Move Eth Transaction DB code to own class and add tx_hash filter
