+++ rotkehlchen/tests/api/test_data_purging.py
@@ -3,6 +3,7 @@ import requests
 
 from rotkehlchen.constants.assets import A_BTC, A_ETH
 from rotkehlchen.db.dbhandler import DBHandler
+from rotkehlchen.db.ethtx import DBEthTx
 from rotkehlchen.db.filtering import ETHTransactionsFilterQuery
 from rotkehlchen.exchanges.data_structures import Trade
 from rotkehlchen.fval import FVal
@@ -97,7 +98,7 @@ def test_purge_single_exchange_data(rotkehlchen_api_server_with_exchanges, added
 
 def test_purge_ethereum_transaction_data(rotkehlchen_api_server):
     rotki = rotkehlchen_api_server.rest_api.rotkehlchen
-    db = rotki.data.db
+    db = DBEthTx(rotki.data.db)
     db.add_ethereum_transactions(
         [EthereumTransaction(
             tx_hash=bytes(),
